started spliiting up structure
