+++ mod/Sequencer/modules/BaseSequenceModule.cpp
@@ -1,0 +1,12 @@
+class CfgVehicles {
+    class Logic;
+    class Module_F : Logic {
+        class ArgumentsBaseUnits {};
+        class ModuleDescription {};
+    };
+
+    class BaseSequenceModule : Module_F {
+        displayName = "Hello World2";
+        category = "MyModuleCategory";
+    };
+};
