+++ mod/Sequencer/modules/HelloWorld/functions/fn_helloWorld.sqf
@@ -1,0 +1,1 @@
+hint "Hello World!";
