+++ mod/Sequencer/sequencer.cpp
@@ -1,0 +1,9 @@
+class CfgFactionClasses {
+    class NO_CATEGORY;
+    class MyModuleCategory : NO_CATEGORY {
+        displayName = "My Custom Modules";
+    };
+};
+
+#include "modules\BaseSequenceModule.cpp"
+#include "modules\HelloWorld\HelloWorld.cpp"
